+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "tsc && node build/index.js",
-    "api2": "tsc && node build/api2.js"
+    "start": "tsc && node build/api2.js"
   },
   "keywords": [],
   "author": "",
