+++ src/interfaces.ts
@@ -1,22 +1,20 @@
 // interfaces.ts
 // this may be dangerous as interfaces will not be able to error during runtime
-export{};
-
 export interface Metadata {
     relatedTransactionId?: number;
     deviceId?: string;
   }
 
 export interface Transaction {
-    transactionId: number;
-    authorizationCode: string;
-    transactionDate: string;
-    customerId: number;
-    transactionType: string;
-    transactionStatus: string;
-    description: string;
-    amount: number;
-    metadata: Metadata;
+  transactionId: number;
+  authorizationCode: string;
+  transactionDate: string;
+  customerId: number;
+  transactionType: string;
+  transactionStatus: string;
+  description: string;
+  amount: number;
+  metadata: Metadata;
 }
 
 export interface Timeline {
@@ -39,4 +37,22 @@ export interface TransactionFin {
 
 export interface RootObject {
   transactions: TransactionFin[];
+}
+
+
+// interfaces for api2
+
+export interface P2PData {
+  transactionDate: string;
+  transactionType: string;
+  transactionAmount: number;
+}
+
+export interface RelatedCustomer {
+  relatedCustomerId: number;
+  relationType: string;
+}
+
+export interface RootRelated {
+  relatedTransactions: RelatedCustomer[];
 }
